Extracted gql & added invite emails for ConfirmTeamCreation
